Merge ExtensionProperties and Extension #27
Done. I keep both classes, because ExtensionProperties is used for determining the original extension path when creating the hybris fake structure by reading the created metadata properties file of the extension and at this point of execution the function which generates the hybris fake structure does not have to know which dependencies an extension have. When instanciating Extension the dependencies must be known, but resolving dependencies at this point would not worth the effort.
But for good coding style I have changed ExtensionProperties in that way that it inherits its properties from Extension and have its own constructor only taking the extension base directory as parameter and setting all other values to null.