start hac-web in a webcontainer #44
I have added a function for determining transitive dependencies for a given dependency to resolve a  bug in creating the pom.xml with custom dependency versions. But there is still a bug leading to the smallest dependency version of a dependency group to be putted into the pom.xml.
Maybe this is the reason why the jetty server reports a missing field in StaticLoggerBinder when trying to start hac-web.